UTF-8 source code. Refactoring of xrRender and build system * UPD: Most of source code converted to UTF-8. Added "/source-charset:utf-8" globally * FIX: Disable warnings in several projects for all configurations (not only for release) * FIX: Enable /MP compile flag globally (F-Ray.props). Remove /MP from projects files * DEL: Unessessary variables in xrRender, xrCore, xrEngine

+++ code/engine.vc2008/xrRender/xrRender/r__dsgraph_render_lods.cpp
@@ -38,7 +38,7 @@ void R_dsgraph_structure::r_dsgraph_render_lods(bool _setup_zb, bool _clear)
 		const size_t iBatchSize = std::min(lstLODs.size() - i, uiImpostersFit);
 		int cur_count = 0;
 		u32 vOffset;
-		FLOD::_hw* V = (FLOD::_hw*)RCache.Vertex.Lock(iBatchSize*uiVertexPerImposter, firstV->geom->vb_stride, vOffset);
+		FLOD::_hw* V = (FLOD::_hw*)RCache.Vertex.Lock(u32(iBatchSize) * uiVertexPerImposter, firstV->geom->vb_stride, vOffset);
 
 		for (size_t j = 0; j < iBatchSize; ++j, ++i)
 		{
@@ -103,7 +103,7 @@ void R_dsgraph_structure::r_dsgraph_render_lods(bool _setup_zb, bool _clear)
 			}
 		}
 		lstLODgroups.push_back(cur_count);
-		RCache.Vertex.Unlock(iBatchSize*uiVertexPerImposter, firstV->geom->vb_stride);
+		RCache.Vertex.Unlock(u32(iBatchSize) * uiVertexPerImposter, firstV->geom->vb_stride);
 
 		// *** Render
 		RCache.set_xform_world(Fidentity);
